requirements.txt

mysql-replication==0.19
PyMySQL==0.9.3
